+++ src/main/java/com/example/TwitterProject/User/UserRepository.java
@@ -2,6 +2,6 @@ package com.example.TwitterProject.User;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserAccount, Long> {
+public interface UserRepository extends JpaRepository<UserAccount, Integer> {
     UserAccount findByEmail(String email);
 }
